This project is a simple book tracker web application that allows users to sign up, log in and track their reading progress
This project is a simple web application that allows users to sign up, log in, and add books they are reading to their personal book list.

Upon login, the user is redirected to a personalized user page where they can add books they are reading, and view the books they have already added.

The application uses Flask as a web framework and MySQL as a database management system. It also includes HTML templates and CSS for styling.